requirements_clean.txt
Flask==3.1.2
flask-cors>=6.0.1
Flask-Login==0.6.3
pushbullet.py==0.12.0
google-generativeai==0.8.5
google-cloud-firestore==2.21.0
python-dotenv==1.1.1
gunicorn
textblob
